Make the free node list a queue (FIFO) instead of a stack (LIFO)
The newly free'd node will be appended to the end of the list, and
poisoned, so it will be the last one to be reallocated and unpoisoned,
which should improve the effective time of the poison in this memory
region. Hopefully this could help discover stale pointers.

Thanks to hostilefork@stackoverflow for pointing this out.