Fix non-ANSI-compliant overlong Native_Specs
In ANSI-C when you turn on --pedantic, it informs you that there is a
maximum length of string literal that a compiler has to support.  The
number is weird: 509 characters, ostensibly for 512 and CR + LF + '\0'

    http://stackoverflow.com/questions/19365391/

This fixes that issue by encoding it as an array of hex bytes, which
silences the warning.  For simplicity and clarity, I also moved the
array's redundant length check into a separate variable.  (Note this
is triply-redundant: values from the generation process, the size
check Rebol adds onto the zlib data, and the size kept by zlib itself.)

But beyond that: stepping into the code starts with decompressing
the native specs.  It was already breaking rules: putting const
data into a fake series (and a series considers itself the owner
and freer of its data!)  This overdue simplification of the compression
This simplification gets rid of an unnecessary mutability cast putting
a const* into a temporary series' owned data.