Test NONE in sys-start.r before triggering attempt
Something that strongly suggests to me that Rebol wasn't loaded up
under a debugger all that often is that there are two triggers in
sys-start.r where a failed `attempt` is likely, causing a Throw_Error
to happen.  If you've set a breakpoint on Throw_Error, you will
always be hitting those twice...which is distracting.  I'd think anyone
who was running under a debugger would get in there and fix that (it
was one of the early things I did, patching it in for Coherence 1.5).