Patches for turning on -Wtype-limits
These are just some basic fixes that come from the compiler noticing
when you have an unsigned type and test if it's less than zero, or
other things of that sort.  It's a good check to have on!

Note the weird situation with WAIT_OBJECT_0.  The WinApi has
effectively defined another name for 0, which means you may be
tempted to make sure your result is in a range starting with the
value of WAIT_OBJECT_0.  But now you're testing for something being
greater than zero...which an unsigned value of DWORD always will be.
I took out the check, but added an assert that WAIT_OBJECT_0==0
in case it ever changes (it probably won't, but...)