non-ROUTINE! patches for enabling -Wsign-compare
Signed and unsigned comparison rules are kind of annoying, but they
are important to know.  Quick quiz: what does this output?

    #include <stdio.h>
    int main() {
        unsigned int e = 304;
        int b = -1020;
        if (e + b > 1020) puts(">1020"); else puts("<=1020");
    }

The sum will be driven to be an unsigned value, since one addends
was unsigned.  And an unsigned value can't hold a negative number.
So it's good to be aware of these.  Sometimes the answer is that
you didn't need to mix after all, sometimes it just needs a cast
to make it explicit and call attention.

This is all the changes that weren't in ROUTINE!/STRUCT! so they
could be considered separately.