Patches for turning on -Wunused-function
There's a lot of dead code in Rebol, and it clutters things up while
reading.  I can say with a fair amount of confidence that there's
nothing particularly valuable in there that could not be rewritten if
it were needed.

So with a focus on the future, I've deleted the dead code (with the
exception of going ahead and finding out what the terminal code for
"clear to end of line is" and plugging that in.)