Add RL_Encode_UTF8, RL_Length_As_UTF8
This is necessary for -Wimplicit-function-declaration, because the linux
host-lib is not part of core...yet it was calling the Encode_UTF8 and
Length_As_UTF8 routines.  Because everything is getting compiled into
the same executable, the Rebol core functions managed to link even
though there was no forward declaration.  But it should have been a
clue that a separate API entry point was needed.