Zero warnings with -Wimplicit-function-declaration
These are the last patches to take care of the implicit function
declaration warning.  Implicit declarations are of course not good,
because you don't know when you're doing something that will fail
later at link time...and you don't get type checking...and it
assumes things return an integer...etc.