No non-constant runtime aggregate initialization
That's a mouthful, but it simply means that if you try to initialize
something with the 'FOO foo = {(...), (...);' syntax in C89, then
all the elements must be constant.  If not, it's illegal.  C99 can
do it and so can C++, but C89 can't and we're actually still thinking
it's useful to build the core with a C89 compiler.