Patches for enabling -Wdeclaration-after-statement
Of all the C89-isms that one has to put up with that bother me, I
think the lack of the ability to just declare things anywhere (other
than just at the top of a block scope) is probably the biggest.  I've
come to a certain kind of peace about it.  But one thing is certain:
if we're going to obey the convention, it should at least be enforced,
and that's what -Wdeclaration-after-statement will do (even if not
building as C89, where you will get it as an error).