Discern extension errors and evaluation categories
Both errors and evaluation categories use RXE_ prefixes in extensions,
which may well have been an oversight.  In part because errors are API
facing, and in part due to my psychic powers that know evaluation
categories are going to be disappearing...I gave RXE_ to the errors and
RXX_ to the evaluation categories.