Touch up the dirent.d_type workaround
Among my few contributions to Rebol source--during the short period of
time when pull requests were being processed--was enough of a patch to
get it building under HaikuOS.  There weren't all these warnings and
processes involved, so I didn't even notice that I had copy/pasted a
piece of code that used variable-length-arrays, a C99 extension.

GCC and Clang will build it, but once you turn on -Wvla-extension they
will mention that it's not C89.  In any case, now that we know more I
fixed it up a bit.  We'll have to make a new Haiku build here, one day.