Fix -Woverlong-strings in host boot code
This reuses the C array generator rewritten for Native_Specs, which I
isolated out into a %common.r file.  I'm not very familiar with what
the host boot is doing, but as mentioned in the Native_Specs commit
c89 string literals can only be 509 characters long (and c99 extends
this only to 4095).