Remove handful of INLINE markers
INLINE is not standardized in C89, and it has weird properties in both
C++ and C99 (although C99 apparently does something slightly better):

https://gustedt.wordpress.com/2010/11/29/myth-and-reality-about-inline-in-c99/

It's really only a hint to the compiler that it is welcome to ignore,
so all you're doing is telling it that it gets to decide anyway.  Hence
for the moment, for the sake of "correctness first" I'm taking the few
INLINE markers out.  After profiling it can be decided if this is
something to try doing...but enough has changed about the world of
compilers to rethink usages anyway.