Update Zlib to modern function protototypes
Long term the right answer for something like this would be to have a
way to call out some library dependency files to be built as C and not
C++ (even when Rebol is being built as C++).  In the meantime I just
edited it (by hand :-/) to the new function prototypes.

Funny, I thought I remembered that zlib could do C++ which is one of
the reasons I'd done the PR from back when I made the subsetting
script for the newer version.  I think what I was remembering was that
it had a conditional zconst you could enable or not, and wires crossed
that I thought it had #ifdefs for modern function prototypes.