Standardize routines that take 'uni' flag
There are several routines whose pattern is to take in a pointer that
may be either a pointer to a REBYTE or a pointer to a REBUNI, and
then a 'uni' flag as a REBOOL tells you which it is.  Traditionally
these took a REBYTE*, and maybe one took a void*.

I decided to switch it to void* and do some basic common-pattern on
the internals of the routines, to produce a 'bp' byte pointer and 'up'
unicode pointer from the void, and use the one that made sense from
the flag.  The compiler will handle the optimization if one is not
used; and having both just makes things clearer so fewer casts are
needed.

There's a technical issue in the sense that it used to have casting
problems in the call which are now fixed, but it will take any type
without complaint.  My perspective on that is that without a cast,
it will be possible in the future to make some kind of C++ type
argument that can only be coerced from REBYTE* or REBUNI* and check
that.  But with the casts, it just is more junky and would make things
muddier when that day comes.  In the meantime, the prototype helps
cue the reader to realize that the type can be either.