Make #define _GNU_SOURCE conditional on C build
I don't completely understand what _GNU_SOURCE does, but without it a
C89 build does not find things like 'environ'.  If you build in C++
it comes up as a duplicate symbol, already included by the compiler.
So I put it conditional under `#ifndef __cplusplus` and it seemed to
work, so that's what it is for now.