skip based on static not "params named scan_state"
In make-headers.r, there was something of an oddity that if you had a
parameter named "scan_state" it wouldn't output a header for it.  The
better way to signal that in the scanner is to declare something as
static, so this switches the code over to use that.  Less quirky.