Separate Host_Lib into its own .inc file
Rebol does not have include guards on its header files, so everything
can be included exactly once.  While revisiting that may be
worthwhile, it does have some advantages.  Yet putting data
definitions in a header...especially one included indirectly under
switches you may not control...is confusing.  And Host_Lib has been
in a header and has been the bane of my linking-to-Rebol-existence.

For everyone who remembers the `Host_Lib` duplicate-symbol-or-
not-defined problems linking RenCpp/RenGarden...this is going to
put an end to those for good.  Yay for making life easier!!!