mark 'extern' and signatures on hostkit routines
The hostkit doesn't use a mechanism like make-headers.r in order to
automatically expose routines from one .c file to another.  Externs must
be specified manually, and several were not.

Additionally, the types need to be correct in C++.  It's good to have
the check, and since these are internal to the host implementation the
fact that they have name mangling isn't a problem (so they don't need
to be extern "C").