Use correct WinApi types (DWORD, LONG, HMODULE..)
The WinApi is pretty well standardized in terms of the types of the
arguments, so one shouldn't be passing a pointer to an 'int' when
it expects a 'DWORD' (for example).  With C++ you get notified
about the bad pointer casts and type issues, so this fixes them for
the Windows host.

Notice this calls out bugs...for instance, there were two error
handling branches in the serial code, and the second (which sent a
signal) would never be called.