Casts for MultiByteToWideChar to get Win32 build
I'm skeptical of some of the casting in here, and don't really want
to go through it with a fine-toothed comb, so I'll leave it for
@ShixinZeng and others to look through.  There's a lot going on,
along with OEM codepages (I read about it being deprecated) and
questions about the sizes of all the buffers...whether the counts are
characters or bytes, etc.

Hopefully this is a start for review to make sure all the types line
up and that it's working properly.  The type-checking should help
under the C++ build.