Eliminate redundancy across code for various hosts
Hostkit was set up to scan exactly one file for function headers to
export as an API.  That file was host-lib.c, and as the number of
platforms grew the "copypasta" grew.  Each POSIX-like host--even if
it were only slightly different--would wind up with a complete copy
of the other POSIX routines.

Given that this was untenable, I went through for an effort to break
out the routines into functional groups...and made a way to indicate
in the %file-base.r which source files you want to run through the
indexing process.  Those files are preceded in the list by a '+'.
Combined with doing the paths relative to os/ and not os/(base-name)
it becomes possible to transclude functions from other OS directories.

Perhaps not super-elegant, but a whole lot better than having all
those slightly-different-because-someone-forgot-to-update-a-place
copies.  Should at least make life a little more bearable pending a
proper PORT!/host model...