Remove debug output from POSIX OS_Call_Process
There was a stray printf that was outputting the return value of the host
code; regardless of error or success.  So if you ever made a CALL it
would say 'ret: 0' if successful.  Note that this was not the return code
of the process (that's written into 'exit_code'), it's the hostkit API status.