Remove faux "to end of line" comment in makefile
I was actually the one who put this comment on the end of the line
during the Atronix split.  But the # feature in makefiles for comments
can only start a line, you can't use it as an end-of-line comment.  So
it was passing a # to the script as an argument.