`make make` working between Win/Linux Makefiles
This is an adjustment of @earl's patch to add the BIN_SUFFIX to the
REBOL_TOOL.  While that's good for sniffing to see if the tool is there,
a Windows-generated makefile will call the REBOL_TOOL r3-make.exe,
so if you are running Linux on that same makefile it won't be able to
retarget it.  By keeping the tool extension-agnostic but making the
detection take the extension into account, we get both options working.