Clean up binary-to-c, make it only accept BINARY!
This routine was added in order to make ANSI c89/c99 compliant arrays
with hexes instead of overlong strings.  But it was willing to take a string
in…by making it take a binary it forces the caller to be clear about what
bytes they want to be in the array.  Cleaned up the function a little too.