Very simple table-with-header processing routines
These functions are not designed for performance and are very rough,
but allow for clearer access to the kinds of data structures used in the
make make and make prep steps.  Some basic routines like this would
be nice in the core if they were simple and not too fancy.  For inspiration:

http://codeconscious.com/rebol-scripts/rowsets.r