Add OS_ID parameters back to make make in makefile
I mistakenly believed that since the OS_ID looked "hardcoded" in the
makefile that it wasn't where the OS_ID was coming from as an
argument…but it turns out that hardcoding was just the default.  I
wanted the default cleared but not to remove the parameter.  Noticed
by @ShixinZeng