.FORCE in makefile.boot, detect BIN_SUFFIX, fixes
This is a few hopeful fixes to issues introduced in the last commit
surrounding trying to change 'make make' to 'make makefile'.  I
still think that's a good idea, but the mechanics of make are getting
in the way; I'll ask on StackOverflow about it.  In the meantime just
go back to 'make make'.  Also does auto-detect on BIN_SUFFIX so
the makefiles can be more generic and cross-platform in a make-make
sense once generated.  (Probably not the last twist we'll see on
that particular issue)