Kill catch-all and ?= (not working on all platforms)
The catch-all isn't working everywhere, so a better solution needs to
be found for that.  ?= has a problem as well - it only defines the
compiler if it's not already defined.  It turns out that some platforms
have a make that defaults CC to 'cc', and then doesn't find gcc.  Both
of these are good features but we need to find a way that works more
uniformly...