Make always_malloc into a PG_Program_Global
When building and linking Rebol against something other than host_main.c,
it's better to have the always_malloc flag be something that is taken care of
with the other pieces of global state vs. having to generate a linkable extern
within each client.  This takes care of that by putting the flag into
sys-globals.h along with making it a debug-build-only feature.