Bugfix + malloc() and free() for buffers in Windows CALL
Originally the code used OS_Make, which may abstractly be not
implemented directly as malloc, and needs to be paired with OS_Free.
Which it was - however, it also called realloc() and there's no
guarantee that however the memory was allocated would be able
to be respond to realloc.  (It so happened everything worked as
Make_Mem was just malloc under the hood, but that may change)
Additionally, I'd screwed up the transformation of the calls into
OS_ALLOC_ARRAY somehow and when a buffer needed to
expand it only allocated to size of char instead of wchar_t.  Both
issues fixed, and a similar issue with a wstrdup potentially
being freed with OS_FREE.