Error handling overhaul, includes CC#1743
(Note QUIT now takes an INTEGER! value only, and has no /NOW
refinement.  A small step for now, but ideally EXIT will be shifted to
act as QUIT/RETURN, and QUIT will be arity zero meaning EXIT 0)

This is a major rewrite of the setjmp/longjmp-based error handling,
and it's not really that reasonable or feasible to break it up much
more than this.  These things are all inter-related.

Central ideas are that there aren't two separate saved states
to worry about (Halt_State is integrated into Saved_State) and the
PUSH_CATCH/DROP_CATCH macros are clearer and simpler,
with less chance for making mistakes in usage.

The mechanism for including a REBVAL associated with an error
has been simplified and formalized a little more.  Though simpler,
an attempt to further tighten the grip on this "sneaky" way of having
one error-per-task turned out to complicate the code a bit too much,
and so it could still be improved with some better asserts.