Comments and get rid of IS_XXX for errors
I specifically targeted IS_RETURN for deletion originally because there
was a new ANY-FUNCTION! class of value called RETURN!, and it was
easy enough to test VAL_ERR_NUM(error) against RE_RETURN to
remove the competition between the tests.  I'll probably not be adding
a new data type in the formal commit, but still it's better to be explicit
about the error number testing, especially given there are so few
instances of that.