Mutability cast workaround for VAL_EVENT_SER +1 const
This was a little confusing until I realized what was happening; if you have a
pointer to a const structure and then take the address of something in that
structure, then that component will be a const address--even if the element
is non-const.  So the trick used to make VAL_EVENT_SER usable as an
LValue messes with the ability to use it on the right hand side.  One could
rethink the macro to split it into two...one for LHS and one for RHS, but
doing this for now.

Also fixes another const issue, with tuples.