Add a stock empty block value
So long as we have a NONE and an UNSET, an empty block can come
in handy as well.  Protect it and lock it, so that it stays empty.  (User can
override protect, but not locking ATM...use PROTECT for the message
and LOCK to make sure even if they UNPROTECT it still stays empty).