DS_SKIP replaced with DS_PUSH calls, incl. TRASH!
"Skip" is not a typical stack operation, and DS_SKIP was used to do
shady things like do a DS_POP or DS_DROP and then get the value back.
This replaces combinations of DS_SKIP+DS_SET to an appropriate
DS_PUSH_XXX variant, adding the opportunity to say you want to push
an uninitialized value as either "safe" or "unsafe" TRASH!

The idea of TRASH! is to provide both safe and unsafe variations of a
value that the user should never see.  A safe variant is that which will
not trip up the garbage collector, and will be an UNSET! in release
builds.  By setting the function output to be a safe TRASH! we are
able to catch cases where a function dispatcher does not consciously
set the output value for that dispatch.  (If we use UNSET!, there is
no way to tell if it wrote unset or just left the one we put there at
initialization.)