Change functions taking 'REBVAL*ds' to take Call struct
In commit 23f6003, I mentioned that the implicit variable "ds" that
was passed to natives and actions could sometimes refer to the
first cell of the call stack frame, and if it was passed around be
replaced with D_OUT.  There were also other cases which intended
to be passing the entire state of the call as an object.  For these
cases I now introduce a structure to represent that, Reb_Call, which
is passed instead.  This distinguishes the case and makes it clear
that what is being passed is not a value nor conceptually an ordered
series of values.

At the moment, it just contains the DSF index in effect at the time
of the call for simplicity.  It will be changed shortly.

So long as I was changing the prototypes for all the functions, I
figured that typing the return value to indicate it expected to return
one of the R_XXX enum values would be good to do also.  To
try and keep the interfaces on standard types, we don't use an
actual enum because those have implementation-defined types
in C.