Eliminate the "Top of Stack Plus One" Volatile Protocol
There was a bad habit in the Rebol codebase to perform a "DROP" of
the data stack (a POP without making a copy of the value)...and then pass
around the pointer to the slot where the popped value was.  Any operation
that pushed values to the stack could invalidate the value at that address,
so there were warnings and disclaimers related to this like "use the value
quickly..." or "stack volatile!"

The main motivation seems to have been in order to keep the recipient
of certain pointers from a routine from "having to worry" about doing a
pop of the value.  This is a false benefit, because they now have to worry
about something trickier and crashier...and it ties the hands of the stack
abstraction to do things like writing debug garbage into freed values
to catch people using memory they shouldn't.

So this gets rid of TOS1 and DS_NEXT, laying the groundwork for some
further important changes...