Make assertion use SERIES_TAIL and not SERIES_LEN
It's a bad thing that in C where "strlen" does not include the null
terminator in the count, and that the notion of the length of something
is how many data elements it has, that SERIES_LEN includes
the terminator.  Hence this assert was always true.

SERIES_TAIL is much less obvious and those macros need to be
fixed, but that's another commit.