Eliminate hidden dependency of DS_OUT
DS_OUT was a global macro for accessing "the output".  This is a hidden
dependency which some functions would use just as scratch space to
write in, while others were actually intending to write an output.  This uses
the D_OUT macro instead (which captures the explicit stack parameter)
in those cases where it fit, and parameterized routines that needed it.
(or fetched directly from DSF in function dispatchers)

(The other DS_* macros will be getting similar treatments...)