Balance stack for negative?, positive?, zero?
Rather than just make a local variable for the zero comparison, these
routines would manually increment the stack (and then refer to the
incremented stack position as D_ARG(2)).  This led to an unbalanced
stack at the end of the operation.  Since use of the data stack is not
necessary, this just makes a REBVAL local and uses that.