Use data stack and non-TOS for REDUCE/COMPOSE
This is an effective reversal of the choice in the Atronix build to
move away from using the data stack as the place where both
reduce and compose first evaluate their values, and then copy
them off.  If you try to do it in one pass, then you don't know
how many elements the newly formed series is going to need;
so you'll either allocate one that's too big or run too many
Expand_Series calls as you keep resizing (or both).

The method in the code is effectively the same as was still in
mainline Rebol, although the interface is updated a bit and it
is commented.  It also does not have them return their result
on the top of stack, rather written into a variable of the caller's
choice that may-or-may-not-be-on-the-stack.