Add the balanced stack assertion
With the fixes in to support the check, this makes sure that if a
function call does not trap an error, then each step of the Do_Next
will end with the data stack exactly where it started.

There are certainly other cases of it we'll hit in the wild; this is just
what the tests and building hostilefork.com turned up.  But it prints
out a debug message that should help catch the problems if the
output line is reported.