substitute VAL_FUNC_ARGS => VAL_FUNC_WORDS
There were alerady two names for the same thing.  VAL_FUNC_ARGS is
a bad name for the words list, because not everything in it is a parameter
(some are locals, there's a SELF slot, and the definitionally scoped return
word will go in here too).

Also it is convenient in the code to consider the difference between
"parameter" and "argument", which Wikipedia has decided is a distinction
between "formal arguments" and "actual arguments".  Whatever one thinks
of that, it reads a bit better to have variables called "arg" and "param" than
"formal" and "actual".

Also, the macro already existed.