Allow marking of const REBVAL* in GC routines
The call frame structure has a const REBVAL* for the function, which is
what motivated this.  But it's just good to have const correctness when it
comes along (in its own commit to avoid confusing matters).

This still has that problem with needing the mutability cast for events.
It's explained by the VAL_EVENT_SER macro.