Use DS_DROP_TO for all DSP updates
Trying to isolate commits down, and this was something that could be
put in its own before the "legit series data stack".  Basically, rather than
setting the stack pointer arbitrarily the stack will offer the ability to drop
it to an earlier location, and in debug builds that will poison the stack
locations above that point to help catch accesses to invalid memory.