Make Data Stack an ordinary law-abiding series
While the data stack is a series, it has traditionally done strange (broken)
things with expansion and tracking its size. Pushing a function call would
pick arbitrary numbers like "20" to make space on the stack for what a
native might use before re-entering the Do_Next loop, and then the
pushes themselves would not check to see if expansion was needed...
leading to crashes if 20 wasn't the actual bound. The stack was not
terminated as a series with a REB_END value due to the perceived
cost of keeping it up to date on pushing and popping, and the length was
updated as a separate global variable while not keeping the length in
the series up to date.

There may be performance benefits to this, but changing to use less
stack in general (by way of writing results directly to the target memory
slot) as well as using a stable memory allocation for call frames shifts
the focus. The "deferred terminator" concept could be implemented for
series as a whole, by only terminating them when a pointer is requested
to their data to be iterated as an array (similar to how C++ strings do not
internally terminate unless you ask them for their byte buffer pointer,
at which point it performs the termination).

This change involved reworking the highly questionable "Redo_Func"
dispatch, which now does not try and modify the existing call frame
and arguments but pushes a new frame based on it and then pops it.