Remove Ordinal Natives dispatch hack
The "efficiency" hack of calling to PICK from inside Rebol so that FIRST,
SECOND, etc. do not have to be mezzanine was fairly ugly.  And a very
narrow optimization anyway.  The dedicated call frame stack is more
controlled than just pushing parameters onto the data stack was, so
so a cleaner method of this would be necessary.

But probably the best way to look at it is to figure out how to make
system-wide speed improvements vs hacks just applying to one case.