Workaround for NULL warning in DS_TOP macro
The macro itself is already a workaround because Panic() returns a
void and we need a pointer in an expression-like slot, so NULL gives
us a (likely) crash.  But in the C++ build this needs to be nullptr to work:

http://stackoverflow.com/a/27951801/211160

More generally, NULL should be defined to nullptr in Rebol to check
for type mistakes.  But that's beyond the scope of this branch.