Rename VOID macro to NOOP for Win32 API compatibility
Apparently Windows defines VOID for the C void type, which is quite
superfluous.  I thought it would be a good name to use for a void
value as "NOOP" looks kind of weird, and it's also more explicit
that it's not a `do {} while (0)` under the hood, or similar.  But... noop.