Get rid of EVAL_TYPE
There was a level of indirection on the processing of values in the Do
loop where each datatype was given an "evaluation type".  Most things
had an eval_type of "Self" indicating that when the evaluator hit them
that they should just be their own value, but active types like WORD
or PAREN were given ET_WORD or ET_PAREN.

Generally this just confused things, for instance that ET_PATH covered
path!, get-path!, and set-path! but not lit-path!.  And ET_FUNCTION
covered all the function types except OP!.  Really a switch on the
actual type in the Do loop is much better for clarity, and the only argument
could be that using the eval types makes it easier to get the values in
the switch sequential so that there's a jump table optimization from the
compiler on the switch.  That's easy enough to do with the REB_*
types as well, and even cleaner if they're ordered to line up with all
the "self" evaluating types after all the active types.

This removed the ET_* from c-do.c, as well as the generation of the
Eval_Type_Map via make-boot.r... and put the headers for the columns
in as code to types.r using the common.r routine used for w/systems.r